Reverted some of the refactoring
